django2/requirements.txt
asgiref==3.8.1
common==0.1.2
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
future==1.0.0
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
serializers==0.2.4
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.1
utils==1.0.2
Werkzeug==3.1.3